+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,8 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.CardSymbol;
-import blackjack.domain.card.CardValue;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
