+++ src/test/java/blackjack/fixture/CardFixture.java
@@ -1,6 +1,6 @@
 package blackjack.fixture;
 
-import blackjack.domain.Deck;
+import blackjack.domain.card.Deck;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardSymbol;
 import blackjack.domain.card.CardValue;
